+++ src/objtools/blast/gene_info_reader/gene_info_reader.cpp
@@ -240,7 +240,7 @@ static bool s_GetMemFilePtrAndLength(CMemoryFile* pMemFile,
 {
     if (pMemFile != 0)
     {
-        nRecs = pMemFile->GetSize() / (sizeof(TRecordType));
+        nRecs = static_cast<int>(pMemFile->GetSize() / (sizeof(TRecordType)));
         if (nRecs > 0)
         {
             pRecs = (TRecordType*)(pMemFile->GetPtr());
